backend/agents/requirements.txt
google-adk[extensions]>=1.28.0
fastapi>=0.115.0
uvicorn>=0.34.0
httpx>=0.27.0
